fix(nonce): add nonce from main
