+++ index.html
@@ -16,7 +16,7 @@
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
     <link rel="manifest" href="/manifest.json" />
-    <style>
+    <style nonce="{{ styleNonce }}">
       body {
         margin: 0;
       }
@@ -124,7 +124,7 @@
           </svg>
         </div>
       </div>
-      <script>
+      <script nonce="{{ styleNonce }}">
         particlesJS('particles', {
           particles: {
             number: {
