+++ src/shared/providers/BasicProviders.tsx
@@ -5,11 +5,22 @@ import { Theme } from 'src/core/theme'
 import { InnerI18nProvider } from './InnerI18nProvider'
 import { i18n } from './i18n'
 
-export const BasicProviders = ({ children }: PropsWithChildren) => {
+// eslint-disable-next-line no-restricted-imports
+import createCache from '@emotion/cache'
+
+export const BasicProviders = ({ children, nonce }: PropsWithChildren<{ nonce?: string }>) => {
+  const emotionCache = nonce
+    ? createCache({
+        key: `fix-nonce`,
+        prepend: true,
+        nonce,
+      })
+    : undefined
+
   return (
     <I18nProvider i18n={i18n}>
       <InnerI18nProvider>
-        <Theme>
+        <Theme emotionCache={emotionCache}>
           <SnackbarProvider>{children}</SnackbarProvider>
         </Theme>
       </InnerI18nProvider>
